Fix OSPF area configuration for loopback interface
Currently, the loopback interface of a router is always assigned to OSPF
(v2&3) area 0. On a router without a direct connection to area 0, this
makes addresses assigned to the loopback interface unreachable for other
routers.

This patch makes the existing igp_area link parameter also a node parameter
so that a router can access it and use its value when configuring the
loopback interface.

Signed-off-by: Martin Buck <mb-tmp-tvguho.pbz@gromit.dyndns.org>